Revert catalog changes

+++ src/algos/photometry.h
@@ -21,7 +21,7 @@ typedef struct {
 	float mag;	// visible magnitude (V filter), for sorting and debug
 	float BV;	// B magnitude - V magnitude, -99.9 if not available
 	float teff; // Gaia Teff
-	int index; // Order in the Gaia results table. This is used to match
+	uint64_t index; // Order in the Gaia results table. This is used to match
 					// HDUs in the FITS, in case of excluded stars
 } pcc_star;
